requirements.txt
cycler==0.10.0
decorator==4.4.0
filterpy==1.4.5
imageio==2.5.0
joblib==0.13.2
kiwisolver==1.1.0
llvmlite==0.29.0
matplotlib==3.1.0
networkx==2.3
numba==0.44.0
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.1
